+++ services/madoc-ts/src/frontend/shared/capture-models/_components/ViewDocument/components/ViewSelector.tsx
@@ -19,19 +19,54 @@ export const ViewSelector: React.FC<{
   const { data: service } = useImageService() as { data?: ImageService };
   const croppedRegion = useCroppedRegion();
   const [image, setImage] = useState('');
+  const [mask, setMask] = useState<any | null>(null);
   const selectorId = selector?.id;
 
   useEffect(() => {
-    if (selector && service && selector.state) {
-      const cropped = croppedRegion(selector.state);
-      if (cropped) {
-        setImage(cropped);
+    if (!selector) return;
+
+    if (selector.type === 'polygon-selector' && selector.state && selector.state.shape) {
+      const points = selector.state.shape.points as Array<[number, number]>;
+      if (points && points.length > 2) {
+        const x1 = Math.min(...points.map(p => p[0]));
+        const y1 = Math.min(...points.map(p => p[1]));
+        const x2 = Math.max(...points.map(p => p[0]));
+        const y2 = Math.max(...points.map(p => p[1]));
+        const cropped = croppedRegion({ x: x1, y: y1, width: x2 - x1, height: y2 - y1 });
+        if (cropped) {
+          const ratio = 1;
+
+          setImage(cropped);
+          const Shape = selector.state.shape.open ? 'polyline' : 'polygon';
+          setMask(
+            <svg width="100%" height="100%" viewBox={`0 0 ${(x2 - x1) * ratio} ${(y2 - y1) * ratio}`}>
+              <defs>
+                <mask id={`selector-mask-${selector.id}`}>
+                  <Shape
+                    points={points.map(p => `${(p[0] - x1) * ratio},${(p[1] - y1) * ratio}`).join(' ')}
+                    style={{ fill: 'white', stroke: 'white', strokeWidth: 2 }}
+                  />
+                </mask>
+              </defs>
+              <image href={cropped} mask={`url(#selector-mask-${selector.id})`} width="100%" height="100%" />
+            </svg>
+          );
+        }
       }
     }
-    if (selector && selector.revisedBy && selector.revisedBy[0]) {
-      const cropped = croppedRegion(selector.revisedBy[0].state);
-      if (cropped) {
-        setImage(cropped);
+
+    if (selector.type === 'box-selector') {
+      if (service && selector.state) {
+        const cropped = croppedRegion(selector.state);
+        if (cropped) {
+          setImage(cropped);
+        }
+      }
+      if (selector.revisedBy && selector.revisedBy[0]) {
+        const cropped = croppedRegion(selector.revisedBy[0].state);
+        if (cropped) {
+          setImage(cropped);
+        }
       }
     }
   }, [croppedRegion, selector, service]);
@@ -42,7 +77,7 @@ export const ViewSelector: React.FC<{
 
   return (
     <CroppedImage
-      data-size="tiny"
+      data-size="small"
       $fluid={fluidImage}
       style={{ margin: inline ? 'none' : '0 .5em', background: '#f9f9f9' }}
       onClick={e => {
@@ -53,7 +88,11 @@ export const ViewSelector: React.FC<{
         }
       }}
     >
-      <img src={image} data-madoc-id={selectorId} alt="cropped region of image" width={fluidImage ? '100%' : 100} />
+      {mask ? (
+        mask
+      ) : (
+        <img src={image} data-madoc-id={selectorId} alt="cropped region of image" width={fluidImage ? '100%' : 100} />
+      )}
     </CroppedImage>
   );
 };