+++ services/madoc-ts/src/frontend/shared/capture-models/editor/selector-types/PolygonSelector/PolygonSelector.tsx
@@ -1,3 +1,5 @@
+import { InputShape } from 'polygon-editor';
+import { ViewSelector } from '../../../_components/ViewDocument/components/ViewSelector';
 import { BaseSelector } from '../../../types/selector-types';
 
 export interface PolygonSelectorProps extends BaseSelector {
@@ -5,10 +7,27 @@ export interface PolygonSelectorProps extends BaseSelector {
   type: 'polygon-selector';
   hidden?: boolean;
   state: null | {
-    points: Array<{ x: number; y: number }>;
-    orientation?: number;
-    readingOrientation?: number;
-    readingDirection?: string;
-    type: string;
+    type: 'polygon' | 'polyline' | 'line';
+    shape: InputShape;
+    svgPreview?: string;
   };
+}
+
+export function PolygonSelector(props: PolygonSelectorProps) {
+  return (
+    <div>
+      {props.state && props.state.shape && !props.state.shape.open && props.state.shape.points.length > 1 ? (
+        <ViewSelector
+          fluidImage
+          selector={{
+            id: 'test',
+            type: 'polygon-selector',
+            state: props.state,
+          }}
+        />
+      ) : (
+        <>Draw a shape</>
+      )}
+    </div>
+  );
 }